lab4/worker/requirements.txt
pika==1.3.2
Pillow==10.2.0
psycopg2-binary==2.9.9
